Update Link offsets and phrases, add project details
Updated the offset for the 'about', 'projects', and 'contact' Links in the Header component for better site navigation. Made changes to the text on the 'about' and 'projects' sections for clarity. Added individual project details to the 'projects' section instead of the previous 'Under Construction' placeholder. Lowercased the 'r' in 'reserved' in the copyright notice in the Footer component for grammatical correctness.