Adjust offset in Header link and refine padding and content in Main component
The commit adjusts the offset for the 'hello' Link component in the Header.jsx file and modifies the padding as well as alters some content inside the Main component of the index.jsx file. The modified contents include a revised biography paragraph, alongside changes in the arrangement and visibility of details in the about and projects sections.