Update projects section and add 'Under Development' notice.
The layout and details of the projects featured in the projects section have been updated. Changes include the inclusion of a direct link to each project's live page. Moreover, an 'Under Development' notice has been added to the about me section informing viewers that it is still being worked on.