Update Contacts component layout
The Contacts component jsx file was updated with a new layout. An option to send an email was added to the left side, while contact information was placed on the right. The text indication for each side was also swapped. This was done to improve contact functionalities and organization within the page.