Update header and link style in LatestBlogPosts
The heading style of LatestBlogPosts component has been updated to a more interactive design. Furthermore, the arrow in the "Coming soon" section has been removed as it was providing redundant information. Now the header displays the text "Read my blog" with a corresponding link decorated with an appropriate CSS class.