Update styling of various components
The commit adjusted the styling of various components in response to changing design requirements. Changes include removing unnecessary CSS class in 'Header' component, modifying padding and space in 'index' and adjusting layout details in various components for better responsive design and aesthetics. These alterations help in improving the visual appeal and user experience of the application.