Update AboutMe and Projects components
The 'AboutMe' component's tooltip label is updated for 'Tailwind' to 'Tailwind CSS', and 'UI' and 'UX' tooltips are removed. In 'Projects' component, 'Libro' project status is updated from 'Beta' to 'Archived' and the project description is revised to past tense, the 'Highstorm' project image and description are updated, and the 'website logo' source is added.