Update contact info and remove development warnings
This commit updates the contact info and GitHub username in the Contacts.jsx component. It also removes the "under development" warnings related to the 'about' and 'contact' sections from Index.jsx, indicating that these sections are now complete. Additionally, the referenced GitHub account name in the 'projects' section is updated.