Refine website styles and interface
This commit addresses a range of UI improvements, mainly focusing on text color and the use of gradients. Updates were made to the Contacts, Projects, Header, Index, Hello, and Social components, adjusting text size and color, background gradients, and link styles for better readability and user experience. The visual enhancements make the elements more distinctive, thus, refining the overall look and feel of the website. Changes also include correcting the path for "og:image" in the meta tag.