Update Projects component layout and functionality
The code updates introduce a revamped layout for the Projects component, focused on enhancing readability and user interaction. The changes include replacing the 'Divider' component with a 'Tab' layout for project descriptions and moving each project's category(like Mobile App, Website) to the header. Additionally, the "faLink" icon is replaced with the "faArrowUpRightFromSquare" for better visual clarity.