Replace FontAwesome with Tabler Icons and update AboutMe component
The FontAwesome icons were removed from package.json and package-lock.json, and replaced with the Tabler Icons library. In the AboutMe component, user information was replaced with a 'Read my story' button. The aim was to make the UI more straightforward by reducing the information shown in favor of a call-to-action button.