Refactor components and switch to Tabler icons
The commit includes a refactoring of several components, aiming to simplify the code and improve readability. Moreover, it switches from using FontAwesome icons to Tabler icons across all instances. This should reduce the project's dependency load and improve the overall aesthetics of the project. Certain SVG properties have been modified as well to calibrate their appearance in the new setup.