Update Header, AboutMe, Footer, and Hello components
The changes include removing the Sometype_Mono font from the Header component, disabling the about, projects, and contact buttons, and adjusting the offset. The AboutMe section was majorly simplified and will be reworked in the future. The Footer text was removed from a Card and its opacity was slightly lowered. The transition in the Hello component was also slightly adjusted.